+++ src/data/OpWaveform.h
@@ -5,9 +5,9 @@
 
 namespace wcopreco {
 
-  class OpWaveform : public std::vector<short> {
+  class OpWaveform : public std::vector<double> {
   public:
-    OpWaveform(int ChanNum, double time_from_trig, int typ, const std::vector<short>& wfm );
+    OpWaveform(int ChanNum, double time_from_trig, int typ, const std::vector<double>& wfm );
 
     OpWaveform(int ChanNum, double time_from_trig, int typ, const int nelems );
     virtual ~OpWaveform() {};
